refactor get_filter_coeff() and get_vend_filter_EK80(), now both in ek80_complex.py
